server/requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
colorama==0.4.4
Django==3.2.13
django-3-jet==1.0.8
django-admin-honeypot==1.1.0
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
importlib-metadata==4.11.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.2.0
zipp==3.8.0
